requirements.txt
Flask==3.1.0
Flask_Cors==4.0.0
pymongo==4.7.2
python-dotenv==1.0.1
